test: Tweak code
